requirements.txt
alembic==1.0.0
aniso8601==3.0.2
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
Flask==1.0.2
flask-marshmallow==0.9.0
Flask-Migrate==2.2.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gevent==1.3.6
greenlet==0.4.15
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.15.4
marshmallow-jsonapi==0.20.2
marshmallow-sqlalchemy==0.14.1
mysqlclient==1.3.13
pika==0.12.0
pipenv==2018.7.1
pkg-resources==0.0.0
psycopg2==2.7.5
pycparser==2.19
pycryptodome==3.6.6
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.5
rabbitmq==0.2.0
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.11
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
Werkzeug==0.14.1
